+++ cms/src/routes/_protected/_dashboard.tsx
@@ -1,0 +1,15 @@
+import { createFileRoute, Outlet } from "@tanstack/react-router";
+import DashboardNavBar from "../../components/dashboardNavBar/DashboardNavBar";
+
+export const Route = createFileRoute("/_protected/_dashboard")({
+  component: RouteComponent,
+});
+
+function RouteComponent() {
+  return (
+    <>
+      <DashboardNavBar />
+      <Outlet />
+    </>
+  );
+}